Assignments API

This page documents the Assignments API of Canvas, see https://canvas.instructure.com/doc/api/assignments for details.

Canvas.assignment

Canvas.assignment_override

Canvas.assignment_overrides

Canvas.assignments

Canvas.create_assignment

Canvas.create_assignment_override

Canvas.delete_assignment

Canvas.delete_assignment_override

Canvas.update_assignment

Canvas.update_assignment_override

Canvas.assignments — Function

Canvas.assignments(c::Course; kwargs...) -> Vector{Assignment}, page_data

Return the paginated list of assignments for the course. Return a vector of Assignments and a dictionary with page data.

Canvas API documentation: List assignments (GET /api/v1/courses/:course_id/assignments)

Canvas.assignments(c::Course, a::AssignmentGroup; kwargs...) -> Vector{Assignment}, page_data

Return the paginated list of assignments for the course and assignment group. Return a vector of Assignments and a dictionary with page data.

Canvas API documentation: List assignments (GET /api/v1/courses/:course_id/assignment_groups/:assignment_group_id/assignments)

Canvas.assignments(u::User, c::Course; kwargs...) -> Vector{Assignment}, page_data

Return the paginated list of assignments for the specified user and course. Return a vector of Assignments and a dictionary with page data.

Canvas API documentation: List assignments for user(GET /api/v1/users/:user_id/courses/:course_id/assignments)

Canvas.assignment — Function

Canvas.assignment(c::Course, a::Assignment; kwargs...) -> Assignment

Return the assignment given by id. Return an Assignment.

Canvas API documentation: Get a single assignment (GET /api/v1/courses/:course_id/assignments/:id)

Canvas.create_assignment — Function

Canvas.create_assignment(c::Course, a::Assignment; kwargs...) -> Assignment

Create a new assignment for the course and return the details. Return an Assignment.

Canvas API documentation: Create an assignment (POST /api/v1/courses/:course_id/assignments)

Canvas.update_assignment — Function

Canvas.update_assignment(c::Course, a::Assignment; kwargs...) -> Assignment

Modify an existing assignment and return the new details. Return an Assignment.

Canvas API documentation: Edit an assignment (PUT /api/v1/courses/:course_id/assignments/:id)

Canvas.delete_assignment — Function

Canvas.delete_assignment(c::Course, a::Assignment; kwargs...) -> Assignment

Delete the given assignment and return the former details. Return an Assignment.

Canvas API documentation: Delete an assignment (DELETE /api/v1/courses/:course_id/assignments/:id)

Canvas.assignment_overrides — Function

Canvas.assignment_overrides(c::Course, a::Assignment; kwargs...) -> Vector{AssignmentOverride}, page_data

Return the paginated list of overrides for this course and assignment. Return a vector of AssignmentOverrides and a dictionary with page data.

Canvas API documentation: List assignment overrides (GET /api/v1/courses/:course_id/assignments/:assignment_id/overrides)

Canvas.assignment_override — Function

Canvas.assignment_override(c::Course, a::Assignment, o::AssignmentOverride; kwargs...) -> AssignmentOverride

Return the details of override specified by id. Return an AssignmentOverride.

Canvas API documentation: Get a single assignment overrides (GET /api/v1/courses/:course_id/assignments/:assignment_id/overrides/:id)

Canvas.assignment_override(co::Union{Group,Section}, a::Assignment; kwargs...) -> AssignmentOverride

Return the override for the given group or section associated with the assignment (404 otherwise). Return an AssignmentOverride.

Canvas API documentation:

Redirect to the assignment override for a group (GET /api/v1/groups/:group_id/assignments/:assignment_id/override)

Redirect to the assignment override for a section (GET /api/v1/sections/:course_section_id/assignments/:assignment_id/override)

Canvas.create_assignment_override — Function

Canvas.create_assignment_override(c::Course, a::Assignment; kwargs...) -> AssignmentOverride

Create an assignment override. One of student_ids, group_id or course_section_id must be present, see API specification for details. Return the AssignmentOverride.

Canvas API documentation: Create an assignment override (POST /api/v1/courses/:course_id/assignments/:assignment_id/overrides)

Canvas.update_assignment_override — Function

Canvas.update_assignment_override(c::Course, a::Assignment, o::AssignmentOverride; kwargs...) -> AssignmentOverride

Update the assignment override. All current overriden values must be supplied if the should be retained, see API specification for details. Return the AssignmentOverride.

Canvas API documentation: Update an assignment override (PUT /api/v1/courses/:course_id/assignments/:assignment_id/overrides/:id)

Canvas.delete_assignment_override — Function

Canvas.delete_assignment_override(c::Course, a::Assignment, o::AssignmentOverride; kwargs...) -> AssignmentOverride

Delete and override and return the former details. Return an AssignmentOverride.

Canvas API documentation: Update an assignment override (DELETE /api/v1/courses/:course_id/assignments/:assignment_id/overrides/:id)